WIP: try to expose to global in separate file
